[puppeteer-service] move schemas to new file
