+++ api/config.js
@@ -1,3 +1,5 @@
-exports.config = {
-  port: process.env.PORT || 3000,
+module.exports = {
+  config: {
+    port: process.env.PORT || 3000,
+  }
 }
