+++ api/schemas.js
@@ -1,0 +1,15 @@
+const htmlPdfRequest = {
+  schema: {
+    body: {
+      type: 'object',
+      required: ['html'],
+      properties: {
+        html: { type: 'string' },
+      }
+    }
+  }
+}
+
+module.exports = {
+  htmlPdfRequest
+}
